feat: compare order of ternary and jumps based on what ends first
This catches cases where we have two jumps in the same block, one for a continue or break out of a loop, and one for an if condition to jump to the after block. In this case we always want to jump to the while condition. Checking the span where the statement was written helps us as the span of the simple continue or break keyword is always smaller than the whole if block, so this is the statement we jump to.